requirements.txt
pygame>=2.1.0
neat-python>=0.92
numpy>=1.21.0
